Sully boolinator's perfection with shortcuts.
These methods are entirely unnecessary, but *some* have chosen to take
exception to their absence.  However, I am not so proud in my absolutely
flawless design that I shall refuse to add a trifling convenience for
those unable to grasp my interface's minimalistic purity in their blind
grasping of instant gratification.

In truth, it is *society* that is to blame.